Refine query check for DDL commands
The method to check if a query is a Data Definition Language (DDL) command has been refined to improve performance and readability. Rather than separately checking if the original query string includes various DDL commands, a new method is used where the original query is converted to lower case and its start is checked against a list of DDL commands.

A new test case was also added in order to ensure if a DDL command was included in a query, but wasn't at the start of the string (as should be in valid SQL commands), it would be properly ignored by the detection method.